Problem statement
Suppose that n people are living in a village has a unique number i(1<=i<=n). It is known that the ages of the people in this village range from 0 to 100.
q interviews are conducted in this village. In response to the question "how many years older are you than person bj?, a person aj answers"-cj years(1<=y<=q). For the case of cj<0, person bj is-cj years younger. For the case cj=0. person bj is the same age.
From the results of the survey determine whether the villages' age assignments are without contradiction. If there are multiple possible age distribution, give the age difference between the eldest and the youngest that is smallest from among the possible age distribution.

Example 1:
Input:
3 2 
1 2 10 
2 3 27
output: 37
Example 2:
Input: 
3 3
1 2 1
2 3 1
3 1 1
output:-1
Example 3:
Input:
3 2 
1 2 100
2 3 100
output: -1