Create Age Fudging
suppose that n people are living in a village. Everyone in this village has a unique number i(1<=i<=n). It is known that the ages of the people in this village range from 0 to 100.
q interviews are conducted in this villages. In response to the question "how manny yeays older are you than person bj?" person aj answers "-cj years(1<=j<=q). For the case of cj<0, person bj is -cj years younger. For the case of cj=0, person bj is the same age.
From the results of the survey, determine whether the villagers' age assignments are without contradiction. If there are multiple possible age distribution give the age difference between the eldest and the youngest that is smallest from among the possible age distributions.
Example 1:
i/p:
3 2
1 2 10
2 3 27
o/p: 37
Example 2:
3 3
1 2 1
2 3 1
3 1 1
o/p:-1